refactor: 추상 controller 패키지 위치 변경
