+++ tomcat/src/main/java/com/techcourse/controller/AbstractController.java
@@ -1,5 +1,6 @@
-package org.apache.catalina;
+package com.techcourse.controller;
 
+import org.apache.catalina.Controller;
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
 import org.apache.coyote.http11.HttpStatus;
