+++ tomcat/src/main/java/com/techcourse/controller/StaticFileController.java
@@ -1,6 +1,5 @@
 package com.techcourse.controller;
 
-import org.apache.catalina.AbstractController;
 import org.apache.coyote.http11.HttpRequest;
 import org.apache.coyote.http11.HttpResponse;
